refactor: Separate case parameters state and functions into seperate hook.
